requirements.txt
locust
cnvrg
requests
selenium
pyslack
Selenium-Screenshot